Switch to ava test runner
