--- test/.eslintrc.yml
@@ -1,8 +1,0 @@
-extends: '../.eslintrc.yml'
-plugins:
-  - mocha
-globals:
-  describe: true
-  it: true
-  expect: true
-  before: true
