--- test/.jshintrc
@@ -1,44 +1,0 @@
-{
-  "globals": {
-    "define": true,
-    "require": true,
-    "describe": true,
-    "it": true,
-    "before": true,
-    "after": true,
-    "beforeEach": true,
-    "afterEach": true
-  },
-
-  "strict": false,
-  "quotmark": true,
-  "node" : true,
-  "es5" : false,
-  "browser" : true,
-  "boss" : false,
-  "curly": false,
-  "debug": false,
-  "devel": false,
-  "eqeqeq": true,
-  "evil": true,
-  "forin": false,
-  "immed": true,
-  "laxbreak": false,
-  "newcap": true,
-  "noarg": true,
-  "noempty": false,
-  "nonew": true,
-  "nomen": false,
-  "onevar": false,
-  "plusplus": false,
-  "regexp": false,
-  "undef": true,
-  "sub": false,
-  "white": false,
-  "eqeqeq": false,
-  "latedef": true,
-  "unused": "vars",
-  "eqnull": true,
-  "sub": true,
-  "expr": true
-}
