+++ test/unwrap.test.js
@@ -1,0 +1,32 @@
+const test = require('ava');
+
+var Model = require('../Model');
+
+var Address = Model.extend({
+  properties: {
+    city: String,
+    state: String
+  }
+});
+
+test('should allow unwrapping', function (t) {
+  var address = new Address({
+    city: 'Durham',
+    state: 'NC'
+  });
+
+  var addressData = Model.unwrap(address);
+  t.is(addressData.city, 'Durham');
+  t.is(addressData.state, 'NC');
+});
+
+test('should allow unwrapping null/undefined', function (t) {
+  t.is(Model.unwrap(null), null);
+  t.is(Model.unwrap(undefined), undefined);
+});
+
+test('should allow unwrapping primitive', function (t) {
+  t.is(Model.unwrap(1), 1);
+  t.is(Model.unwrap('test'), 'test');
+  t.false(Model.unwrap(false));
+});
